+++ src/ska/contrib/django/ska/integration/drf/views/__init__.py
@@ -1,0 +1,1 @@
+from .jwt_token import *